requirements.txt
# V2 - Only what we actually need
google-genai>=1.63.0      # Gemini for generation + grounded search
google-auth>=2.23.0       # Google API authentication
google-api-python-client>=2.108.0  # Google Search Console
openai>=1.12.0            # OpenAI-compatible API (Nebius)
aiohttp==3.10.5           # Async HTTP
python-dotenv==1.0.1      # Environment vars
beautifulsoup4==4.12.3    # HTML parsing
pydantic>=2.9.0           # Data validation (google-genai requires >=2.9)

# Interactive CLI
click>=8.1.0              # CLI framework
rich>=13.7.0              # Terminal output (panels, tables, progress)
questionary>=2.0.1        # Interactive prompts
PyYAML>=6.0.1             # YAML parsing for MDX frontmatter
nest_asyncio>=1.5.6       # Allow nested event loops (questionary compat)